Переработана структура защиты API, теперь все API проверяют accessToken перед выполнением запросов

+++ server/api/tasks/[id].delete.ts
@@ -1,5 +1,6 @@
 import { defineEventHandler } from 'h3';
 import { $fetch } from 'ofetch';
+import jwt from 'jsonwebtoken';
 
 interface Task {
   id: number;
@@ -10,12 +11,26 @@ interface Task {
 }
 
 export default defineEventHandler(async (event) => {
-  const id = Number(event.context.params?.id);
+  const authHeader = event.headers.get('Authorization');
 
-  // Удаляем задачу из JSON Server
-  await $fetch(`http://localhost:3001/tasks/${id}`, {
-    method: 'DELETE',
-  });
+  if (!authHeader || !authHeader.startsWith('Bearer ')) {
+    throw createError({ statusCode: 401, statusMessage: 'Unauthorized' });
+  }
 
-  return { message: 'Задача успешно удалена' };
+  const token = authHeader.split(' ')[1];
+
+  try {
+    const decoded = jwt.verify(token, 'Steblin-Aleksandr-key-124][[') as { userId: number };
+
+    const id = Number(event.context.params?.id);
+
+    // Удаляем задачу из JSON Server
+    await $fetch(`http://localhost:3001/tasks/${id}`, {
+      method: 'DELETE',
+    });
+
+    return { message: 'Задача успешно удалена' };
+  } catch (error) {
+    throw createError({ statusCode: 401, statusMessage: 'Invalid token' });
+  }
 });
